+++ tests/locales/en.php
@@ -1,0 +1,17 @@
+<?php
+
+/**
+ * @author Jared King <j@jaredtking.com>
+ *
+ * @see http://jaredtking.com
+ *
+ * @copyright 2015 Jared King
+ * @license MIT
+ */
+
+return [
+    'phrases' => [
+        'test_phrase' => 'This is a test',
+        'parameter_injection' => 'Testing parameter injection: {{parameter_1}} {{blah}} {{test}}',
+    ],
+];